Rename styles files from index.css to index.module.scss
