+++ react-app/src/components/InputText/index.module.scss
