+++ react-app/src/components/Logo/index.module.scss
